Update eslint config and fix linting errors
