+++ gatsby-config.js
@@ -3,15 +3,15 @@ module.exports = {
     title: 'Algorithm Visualizer',
   },
   plugins: [
-      {
-          resolve: 'gatsby-plugin-google-fonts',
-          options: {
-            fonts: [
-              'Raleway',
-              'Orbitron',
-              'source sans pro\:300,400,400i,700', // you can also specify font weights and styles
-            ],
-          },
-        },
+    {
+      resolve: 'gatsby-plugin-google-fonts',
+      options: {
+        fonts: [
+          'Raleway',
+          'Orbitron',
+          'source sans pro:300,400,400i,700', // you can also specify font weights and styles
+        ],
+      },
+    },
   ],
 }
