+++ package.json
@@ -12,10 +12,12 @@
   "dependencies": {
     "gatsby": "^2.15.9",
     "gatsby-plugin-google-fonts": "^1.0.1",
+    "prop-types": "^15.7.2",
     "react": "^16.9.0",
     "react-dom": "^16.9.0"
   },
   "devDependencies": {
+    "babel-eslint": "^10.0.3",
     "eslint": "^6.4.0",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-react-app": "^5.0.2",
