+++ src/pages/index.jsx
@@ -3,7 +3,7 @@ import Layout from '../components/layout/layout'
 
 const Index = () => (
   <Layout>
-    <p style={{ padding: `20px 20px` }}>This is the home page</p>
+    <p style={{ padding: '20px 20px' }}>This is the home page</p>
   </Layout>
 )
